renaming utills to utils and modifying get variant endpoint
