+++ agenta-backend/agenta_backend/routers/evaluation_router.py
@@ -40,7 +40,7 @@ from agenta_backend.services.evaluation_service import (
     execute_custom_code_evaluation,
 )
 from agenta_backend.services import evaluation_service
-from agenta_backend.utills.common import engine, check_access_to_app
+from agenta_backend.utils.common import engine, check_access_to_app
 from agenta_backend.services.db_manager import query, get_user_object
 from agenta_backend.models.db_models import EvaluationDB, EvaluationScenarioDB
 from agenta_backend.config import settings
