+++ agenta-backend/agenta_backend/routers/testset_router.py
@@ -17,7 +17,7 @@ from agenta_backend.models.api.testset_model import (
     TestSetOutputResponse,
 )
 from agenta_backend.config import settings
-from agenta_backend.utills.common import engine, check_access_to_app
+from agenta_backend.utils.common import engine, check_access_to_app
 from agenta_backend.models.db_models import TestSetDB
 from agenta_backend.services.db_manager import query, get_user_object
 from agenta_backend.services import new_db_manager
