+++ agenta-backend/agenta_backend/routers/variants_router.py
@@ -18,9 +18,9 @@ from agenta_backend.services import (
     new_db_manager,
     new_app_manager,
 )
-from agenta_backend.utills.common import check_access_to_app, get_app_instance
+from agenta_backend.utils.common import check_access_to_app, get_app_instance
 from agenta_backend.models.converters import app_variant_db_to_output
-from agenta_backend.utills.common import check_user_org_access, check_access_to_variant
+from agenta_backend.utils.common import check_user_org_access, check_access_to_variant
 from agenta_backend.models.api.api_models import (
     URI,
     App,
