+++ agenta-backend/agenta_backend/services/evaluation_service.py
@@ -18,7 +18,7 @@ from agenta_backend.models.api.evaluation_model import (
     EvaluationUpdate,
 )
 from agenta_backend.models import converters
-from agenta_backend.utills.common import engine
+from agenta_backend.utils.common import engine
 from agenta_backend.services.db_manager import query, get_user_object
 from agenta_backend.services import new_db_manager
 from agenta_backend.models.db_models import (
@@ -32,7 +32,7 @@ from agenta_backend.models.db_models import (
     CustomEvaluationDB,
 )
 
-from agenta_backend.utills import common
+from agenta_backend.utils import common
 
 from langchain.chains import LLMChain
 from langchain.llms import OpenAI
@@ -663,7 +663,9 @@ async def execute_custom_code_evaluation(
         raise HTTPException(status_code=404, detail="Evaluation not found")
 
     # Check if user has app access
-    access = await common.check_access_to_app(user_org_data=user_org_data, app_id=app_id)
+    access = await common.check_access_to_app(
+        user_org_data=user_org_data, app_id=app_id
+    )
     if not access:
         raise HTTPException(
             status_code=403, detail=f"You do not have access to this app: {app_id}"
@@ -794,7 +796,9 @@ async def fetch_custom_evaluation_names(
     user = await get_user_object(user_org_data["uid"])
 
     # Check if user has app access
-    access = await common.check_access_to_app(user_org_data=user_org_data, app_id=app_id)
+    access = await common.check_access_to_app(
+        user_org_data=user_org_data, app_id=app_id
+    )
     if not access:
         raise HTTPException(
             status_code=403, detail=f"You do not have access to this app: {app_id}"
