+++ agenta-cli/agenta/client/client.py
@@ -121,7 +121,7 @@ def list_variants(app_id: str, host: str) -> List[AppVariant]:
         a list of the variants using the pydantic model
     """
     response = requests.get(
-        f"{host}/{BACKEND_URL_SUFFIX}/app_variant/variants/?app_id={app_id}",
+        f"{host}/{BACKEND_URL_SUFFIX}/app_variant/{app_id}/variants/",
         timeout=600,
     )
